Remove old BGP code
